git-svn-id: svn://192.168.1.10/public/CoffeeMud@5372 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Abilities/Thief/Thief_Palm.java
@@ -42,8 +42,8 @@ public class Thief_Palm extends ThiefSkill
 	private static final String[] triggerStrings = {"PALM"};
 	public String[] triggerStrings(){return triggerStrings;}
 	public int usageType(){return USAGE_MOVEMENT|USAGE_MANA;}
-	public int combatCastingTime(){return 0;}
-	public int castingTime(){return 0;}
+	public double combatCastingTime(){return 0.0;}
+	public double castingTime(){return 0.0;}
 
 	public boolean invoke(MOB mob, Vector commands, Environmental givenTarget, boolean auto, int asLevel)
 	{
